Move business logic to service layer. Closes #89

+++ online-shop-server/src/main/java/com/yashmerino/online/shop/services/ProductServiceImpl.java
@@ -24,14 +24,20 @@ package com.yashmerino.online.shop.services;
  + SOFTWARE.
  +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++*/
 
+import com.yashmerino.online.shop.model.Cart;
+import com.yashmerino.online.shop.model.CartItem;
 import com.yashmerino.online.shop.model.Product;
 import com.yashmerino.online.shop.model.User;
+import com.yashmerino.online.shop.model.dto.ProductDTO;
+import com.yashmerino.online.shop.repositories.CartItemRepository;
 import com.yashmerino.online.shop.repositories.ProductRepository;
 import com.yashmerino.online.shop.repositories.UserRepository;
 import com.yashmerino.online.shop.services.interfaces.ProductService;
-import com.yashmerino.online.shop.utils.Role;
+import com.yashmerino.online.shop.utils.RequestBodyToEntityConverter;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -54,14 +60,21 @@ public class ProductServiceImpl implements ProductService {
     private final UserRepository userRepository;
 
     /**
+     * Cart item repository.
+     */
+    private final CartItemRepository cartItemRepository;
+
+    /**
      * Constructor to inject dependencies.
      *
-     * @param productRepository is the product repository.
-     * @param userRepository    is the user repository.
+     * @param productRepository  is the product repository.
+     * @param userRepository     is the user repository.
+     * @param cartItemRepository is the cart item repository.
      */
-    public ProductServiceImpl(ProductRepository productRepository, UserRepository userRepository) {
+    public ProductServiceImpl(ProductRepository productRepository, UserRepository userRepository, CartItemRepository cartItemRepository) {
         this.productRepository = productRepository;
         this.userRepository = userRepository;
+        this.cartItemRepository = cartItemRepository;
     }
 
     /**
@@ -103,7 +116,13 @@ public class ProductServiceImpl implements ProductService {
      */
     @Override
     public void delete(Long id) {
-        productRepository.deleteById(id);
+        Optional<Product> productOptional = this.getProduct(id);
+
+        if (productOptional.isPresent()) {
+            productRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("Product couldn't be found!");
+        }
     }
 
     /**
@@ -123,6 +142,66 @@ public class ProductServiceImpl implements ProductService {
             return productRepository.getProductsBySellerId(userId);
         } else {
             throw new BadCredentialsException("Bad credentials.");
+        }
+    }
+
+    /**
+     * Add product to the cart.
+     *
+     * @param id       is the product's id.
+     * @param quantity is the quantity.
+     */
+    @Override
+    public void addProductToCart(final Long id, final Integer quantity) {
+        Optional<Product> productOptional = this.getProduct(id);
+
+        if (productOptional.isPresent()) {
+            Product product = productOptional.get();
+
+            UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+            Optional<User> currentUserOptional = userRepository.findByUsername(userDetails.getUsername());
+
+            if (currentUserOptional.isPresent()) {
+                User user = currentUserOptional.get();
+                Cart cart = user.getCart();
+
+                CartItem cartItem = new CartItem();
+                cartItem.setCart(cart);
+                cartItem.setProduct(product);
+                cartItem.setQuantity(quantity);
+                cartItem.setName(product.getName());
+                cartItem.setPrice(product.getPrice());
+                cartItemRepository.save(cartItem);
+
+                product.linkCartItem(cartItem);
+                productRepository.save(product);
+
+            } else {
+                throw new EntityNotFoundException("User couldn't be found!");
+            }
+        } else {
+            throw new EntityNotFoundException("Product couldn't be found!");
+        }
+    }
+
+    /**
+     * Adds a new product.
+     *
+     * @param productDTO is the product DTO.
+     */
+    @Override
+    public void addProduct(ProductDTO productDTO) {
+        Product product = RequestBodyToEntityConverter.convertToProduct(productDTO);
+
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Optional<User> userOptional = userRepository.findByUsername(userDetails.getUsername());
+
+        if (userOptional.isPresent()) {
+            User user = userOptional.get();
+            product.setUser(user);
+            productRepository.save(product);
+        } else {
+            throw new EntityNotFoundException("User couldn't be found!");
         }
     }
 }
